pyproject.toml
[tool.poetry]
name = "testing-de"
version = "0.1.0"
description = ""
authors = ["AnnPenchuk <annapenchuk95@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^7.4.2"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic-settings = "^2.7.0"
httpx = "^0.25.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

